===============================================================================
git-svn-id: file:///home/xapian-svn/xapian/trunk@2444 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/matcher/mergepostlist.h
@@ -59,7 +59,6 @@ class MergePostList : public PostList {
 	virtual PositionList * get_position_list();
 
         MergePostList(std::vector<PostList *> plists_);
-	// FIXME: LocalMatch *matcher_?
         ~MergePostList();
 };
 
@@ -114,7 +113,7 @@ inline bool
 MergePostList::at_end() const
 {
     Assert(current != -1);
-    return current >= plists.size();
+    return (unsigned int)current >= plists.size();
 }
 
 inline std::string
